Added AddParking, ViewParking, fixed Navigation issues.
I added the ability to add parking to the database, as well as keeping track of who's parking it is. You can view your parking in the ViewParkingView, as well as delete them.

I also ran into an issue with the accidental 3 NavigationViews in a row, that was causing issues. Each tab has it's own Navigation now. I spent 3 hours working that out lmao....

ViewParkingDetailsView to come next.