Revised Duplicate Map Annotation Code
I found a bug that would cause the app to duplicate the annotations, regardless of if it was told to or not.

I also moved the code that puts pins on the map from the updateUIView function to the makeUIView function.

I also moved the Coords call from the MapViewContentView to the MapView, and then passed in the variables when the map is needed.